Add a separate version check workflow
I think the reason the publish failed in utility is because we abstracted out two different jobs, and so it may've confused trusted publishing a little bit? I think what needs to happen is, we need to abstract the version check and the publish action separately, then call both in the consuming package.