Get rid of git fetch --unshallow
I don't think we really need it anymore. I initially added it so we can check the package.json version against what it is on main, but now that's not a required step after the docs/versioning refactor. As such, I think we can safely get rid of that step in all workflows.